Renamed classes in the ChessServer\Socket namespace
